Deployment of embedding Recommender using Pywebio/requirements.txt
absl-py==0.9.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
future==0.18.2
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
gunicorn==20.1.0
h5py==2.10.0
hyperopt==0.2.5
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonify==0.5
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.4
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.1.5
Pillow==8.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywebio
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
sklearn==0.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.59.0
urllib3==1.26.4
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
scipy>=0.15.1
scikit-learn==0.22.1
matplotlib>=1.4.3
pandas>=0.19
streamlit
tensorflow
